Implemented signature, #1389 for sync and exportdb
